services/ingestion/requirements.txt
kafka-python-ng==2.2.2
azure-storage-blob==12.24.0
pyarrow==18.1.0
